Small fix on module_lsa (bugging me) 🧹

+++ models/modules_features/module_lsa.py
@@ -9,7 +9,7 @@ import modules_corpora.module_gensim    as module_gensim
 
 # =================================== CONSTANTS DEFINITIONS ===================================
 
-DEFAULT_VALUE : int = 0
+DEFAULT_VALUE : float = 0.0
 NUMBER_OF_WORDS_PER_BAG : int = 5
 
 # =================================== PRIVATE METHODS ===================================
@@ -59,7 +59,7 @@ def lsa_analysis(structure_df: pd.DataFrame) -> pd.DataFrame:
     structure_df['LSA - Embedding per Word Groups'] = structure_df['LSA - Word Groups'].progress_apply(lambda groups_of_words: convert_groups_of_words_to_embeddings(groups_of_words, lsa_model))
     structure_df['LSA - Embedding Groups'] = structure_df['LSA - Embedding per Word Groups'].progress_apply(average_embedding_per_group)
     # LSA Coherence scores
-    structure_df['LSA - First Order Coherence'] = structure_df['LSA - Embedding Groups'].progress_apply(lambda embeddings: compute_coherence_score(embeddings, 1))
-    structure_df['LSA - Second Order Coherence'] = structure_df['LSA - Embedding Groups'].progress_apply(lambda embeddings: compute_coherence_score(embeddings, 2))
+    structure_df['LSA - First Order Coherence'] = structure_df['LSA - Embedding Groups'].progress_apply(lambda embeddings: compute_coherence_score(embeddings, 1)).astype(float)
+    structure_df['LSA - Second Order Coherence'] = structure_df['LSA - Embedding Groups'].progress_apply(lambda embeddings: compute_coherence_score(embeddings, 2)).astype(float)
     
     return structure_df
